fix(rpc): fix the rpc transaction data structure
Fixed the rpc transaction data structure(located in `pkg/rpc/type.go`).
LogMessage's member Namespace and Files added.

Former contains the information of the namespace of the generator.
More specifically, it can be defined by user.

Latter has the information of files to send. MapTable's indexes can be
used as key to find the filename fast. In other words, MapTable's values
are filenames. Next, Indexes is contains the sequence of the files in the
data. It does not have the filename as string but it has filename as
integer which is defined in the MapTable. Sample example is like below.

```
Files.MapTable = ["sample1.txt", "sample2.txt"]
// Timestamp field is ignored
LogInfo = [{"length": 10}, {"length": 20}, {"length": 30}]
Files.Indexes = [0, 1, 0]
// First index means "sample1.txt file data length 10"
// Second index means "sample2.txt file data length 20"
// Third index means "sample3.txt file data length 30"
```

NOTE THAT, You must take care of the Indexes' per type is uint8. That
means, you cannot add the file over 256.